Book Nerd 0.2
Added Main Menu Scene
Added Event LIstener Buttons
Added Level 2 Scene
Fixed Projectile Aim
Created Gamemanager Singleton
Added Game End scene

+++ Assets/Scripts/Player/Projectiles/Projectile.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Projectile : MonoBehaviour
 {
     [SerializeField] protected float speed;
-    [SerializeField] protected float damage;
+    [SerializeField] protected int damage;
     protected Rigidbody2D rb;
     protected float xpseed;
